add IAuthorityFetcher
