--- server/src/main/java/org/apache/iotdb/db/auth/AuthorityFetcher.java
@@ -1,133 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iotdb.db.auth;
-
-import org.apache.iotdb.common.rpc.thrift.TSStatus;
-import org.apache.iotdb.commons.client.IClientManager;
-import org.apache.iotdb.commons.consensus.PartitionRegionId;
-import org.apache.iotdb.confignode.rpc.thrift.TAuthorizerReq;
-import org.apache.iotdb.confignode.rpc.thrift.TAuthorizerResp;
-import org.apache.iotdb.confignode.rpc.thrift.TCheckUserPrivilegesReq;
-import org.apache.iotdb.confignode.rpc.thrift.TLoginReq;
-import org.apache.iotdb.confignode.rpc.thrift.TPermissionInfoResp;
-import org.apache.iotdb.db.client.ConfigNodeClient;
-import org.apache.iotdb.db.client.ConfigNodeInfo;
-import org.apache.iotdb.db.client.DataNodeClientPoolFactory;
-import org.apache.iotdb.db.mpp.plan.execution.config.AuthorizerConfigTask;
-import org.apache.iotdb.db.mpp.plan.execution.config.ConfigTaskResult;
-import org.apache.iotdb.db.qp.logical.sys.AuthorOperator;
-import org.apache.iotdb.rpc.ConfigNodeConnectionException;
-import org.apache.iotdb.rpc.StatementExecutionException;
-import org.apache.iotdb.rpc.TSStatusCode;
-
-import com.google.common.util.concurrent.SettableFuture;
-import org.apache.thrift.TException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.util.List;
-import java.util.Locale;
-
-public class AuthorityFetcher {
-
-  private static final Logger logger = LoggerFactory.getLogger(AuthorizerConfigTask.class);
-
-  private static final IClientManager<PartitionRegionId, ConfigNodeClient> configNodeClientManager =
-      new IClientManager.Factory<PartitionRegionId, ConfigNodeClient>()
-          .createClientManager(new DataNodeClientPoolFactory.ConfigNodeClientPoolFactory());
-
-  public static TPermissionInfoResp checkPath(String username, List<String> allPath, int permission)
-      throws ConfigNodeConnectionException {
-    TCheckUserPrivilegesReq req = new TCheckUserPrivilegesReq(username, allPath, permission);
-    TPermissionInfoResp status = null;
-    try (ConfigNodeClient configNodeClient =
-        configNodeClientManager.borrowClient(ConfigNodeInfo.partitionRegionId)) {
-      // Send request to some API server
-      status = configNodeClient.checkUserPrivileges(req);
-    } catch (TException | IOException e) {
-      throw new ConfigNodeConnectionException("Couldn't connect config node");
-    } finally {
-      if (status == null) {
-        status = new TPermissionInfoResp();
-      }
-    }
-    return status;
-  }
-
-  /** Check the user */
-  public static TPermissionInfoResp checkUser(String username, String password)
-      throws ConfigNodeConnectionException {
-    TLoginReq req = new TLoginReq(username, password);
-    TPermissionInfoResp status = null;
-    try (ConfigNodeClient configNodeClient =
-        configNodeClientManager.borrowClient(ConfigNodeInfo.partitionRegionId)) {
-      // Send request to some API server
-      status = configNodeClient.login(req);
-    } catch (TException | IOException e) {
-      throw new ConfigNodeConnectionException("Couldn't connect config node");
-    } finally {
-      if (status == null) {
-        status = new TPermissionInfoResp();
-      }
-    }
-    return status;
-  }
-
-  public static SettableFuture<ConfigTaskResult> operatePermission(
-      TAuthorizerReq authorizerReq, ConfigNodeClient configNodeClient) {
-    SettableFuture<ConfigTaskResult> future = SettableFuture.create();
-    try {
-      // Send request to some API server
-      TSStatus tsStatus = configNodeClient.operatePermission(authorizerReq);
-      // Get response or throw exception
-      if (TSStatusCode.SUCCESS_STATUS.getStatusCode() != tsStatus.getCode()) {
-        logger.error(
-            "Failed to execute {} in config node, status is {}.",
-            AuthorOperator.AuthorType.values()[authorizerReq.getAuthorType()]
-                .toString()
-                .toLowerCase(Locale.ROOT),
-            tsStatus);
-        future.setException(new StatementExecutionException(tsStatus));
-      } else {
-        future.set(new ConfigTaskResult(TSStatusCode.SUCCESS_STATUS));
-      }
-    } catch (TException e) {
-      logger.error("Failed to connect to config node.");
-      future.setException(e);
-    }
-    // If the action is executed successfully, return the Future.
-    // If your operation is async, you can return the corresponding future directly.
-    return future;
-  }
-
-  public static TAuthorizerResp queryPermission(
-      TAuthorizerReq authorizerReq, ConfigNodeClient configNodeClient) throws TException {
-    TAuthorizerResp authorizerResp = new TAuthorizerResp();
-    try {
-      // Send request to some API server
-      authorizerResp = configNodeClient.queryPermission(authorizerReq);
-      // Get response or throw exception
-    } catch (TException e) {
-      throw new TException(e);
-    }
-    return authorizerResp;
-  }
-}
