+++ server/src/test/java/org/apache/iotdb/db/auth/AuthorizerManagerTest.java
@@ -43,7 +43,7 @@ import java.util.Set;
 
 public class AuthorizerManagerTest {
 
-  AuthorizerManager authorizerManager = AuthorizerManager.getInstance();
+  ClusterAuthorityFetcher authorityFetcher = ClusterAuthorityFetcher.getInstance();
 
   @Test
   public void permissionCacheTest() throws ConfigNodeConnectionException, AuthException {
@@ -92,8 +92,8 @@ public class AuthorizerManagerTest {
     result.setRoleInfo(new HashMap<>());
 
     // User authentication permission without role
-    authorizerManager.getUserCache().put(user.getName(), authorizerManager.cacheUser(result));
-    User user1 = authorizerManager.getUserCache().getIfPresent(user.getName());
+    authorityFetcher.getUserCache().put(user.getName(), authorityFetcher.cacheUser(result));
+    User user1 = authorityFetcher.getUserCache().getIfPresent(user.getName());
     assert user1 != null;
     Assert.assertEquals(user.getName(), user1.getName());
     Assert.assertEquals(user.getPassword(), user1.getPassword());
@@ -102,20 +102,20 @@ public class AuthorizerManagerTest {
     // User has permission
     Assert.assertEquals(
         TSStatusCode.SUCCESS_STATUS.getStatusCode(),
-        authorizerManager
-            .checkPermissionCache(
+        authorityFetcher
+            .checkUserPrivileges(
                 "user", Collections.singletonList("root.ln"), PrivilegeType.CREATE_ROLE.ordinal())
             .getCode());
     // User does not have permission
     Assert.assertEquals(
         TSStatusCode.NO_PERMISSION_ERROR.getStatusCode(),
-        authorizerManager
-            .checkPermissionCache(
+        authorityFetcher
+            .checkUserPrivileges(
                 "user", Collections.singletonList("root.ln"), PrivilegeType.CREATE_USER.ordinal())
             .getCode());
 
     // Authenticate users with roles
-    authorizerManager.invalidateCache(user.getName(), "");
+    authorityFetcher.invalidateCache(user.getName(), "");
     tUserResp.setPrivilegeList(new ArrayList<>());
     tUserResp.setRoleList(user.getRoleList());
 
@@ -133,30 +133,30 @@ public class AuthorizerManagerTest {
       tRoleRespMap.put(role.getName(), tRoleResp);
     }
     result.setRoleInfo(tRoleRespMap);
-    authorizerManager.getUserCache().put(user.getName(), authorizerManager.cacheUser(result));
-    Role role3 = authorizerManager.getRoleCache().getIfPresent(role1.getName());
+    authorityFetcher.getUserCache().put(user.getName(), authorityFetcher.cacheUser(result));
+    Role role3 = authorityFetcher.getRoleCache().getIfPresent(role1.getName());
     Assert.assertEquals(role1.getName(), role3.getName());
     Assert.assertEquals(role1.getPrivilegeList(), role3.getPrivilegeList());
 
     // role has permission
     Assert.assertEquals(
         TSStatusCode.SUCCESS_STATUS.getStatusCode(),
-        authorizerManager
-            .checkPermissionCache(
+        authorityFetcher
+            .checkUserPrivileges(
                 "user", Collections.singletonList("root.ln"), PrivilegeType.CREATE_ROLE.ordinal())
             .getCode());
     // role does not have permission
     Assert.assertEquals(
         TSStatusCode.NO_PERMISSION_ERROR.getStatusCode(),
-        authorizerManager
-            .checkPermissionCache(
+        authorityFetcher
+            .checkUserPrivileges(
                 "user", Collections.singletonList("root.ln"), PrivilegeType.CREATE_USER.ordinal())
             .getCode());
 
-    authorizerManager.invalidateCache(user.getName(), "");
+    authorityFetcher.invalidateCache(user.getName(), "");
 
-    user1 = authorizerManager.getUserCache().getIfPresent(user.getName());
-    role1 = authorizerManager.getRoleCache().getIfPresent(role1.getName());
+    user1 = authorityFetcher.getUserCache().getIfPresent(user.getName());
+    role1 = authorityFetcher.getRoleCache().getIfPresent(role1.getName());
 
     Assert.assertNull(user1);
     Assert.assertNull(role1);
